reorder(lib): introduce ExtractComponentData
`utils/helpers` introduces a new helper type `ExtractComponentData` that
extracts the data type from the constructor of a component. It may help
us avoiding use of `any` when we `mount` or `shallowMount` a component
that has mixins in tests using Vue test utils. As far as I tested,
`mount` and `shallowMount` do not respect data fields in mixins.